requirements.txt
qwen-agent[mcp,gui]
fastmcp
gradio
requests
